requirements.txt
numpy>=1.21.0
pandas>=1.3.0
torch>=2.0.0
PyYAML>=6.0
matplotlib>=3.4.0
plotly>=5.3.0
scikit-learn>=0.24.0
tqdm>=4.62.0
jupyter>=1.0.0 